+++ packages/google_maps_flutter/google_maps_flutter/example/lib/main.dart
@@ -9,6 +9,7 @@ import 'package:google_maps_flutter_android/google_maps_flutter_android.dart';
 import 'package:google_maps_flutter_platform_interface/google_maps_flutter_platform_interface.dart';
 
 import 'animate_camera.dart';
+import 'heatmap.dart';
 import 'lite_mode.dart';
 import 'map_click.dart';
 import 'map_coordinates.dart';
@@ -43,6 +44,7 @@ final List<GoogleMapExampleAppPage> _allPages = <GoogleMapExampleAppPage>[
   const LiteModePage(),
   const TileOverlayPage(),
   const MapIdPage(),
+  const HeatmapPage(),
 ];
 
 /// MapsDemo is the Main Application.