+++ packages/google_maps_flutter/google_maps_flutter_android/android/src/main/java/io/flutter/plugins/googlemaps/Convert.java
@@ -28,6 +28,9 @@ import com.google.android.gms.maps.model.RoundCap;
 import com.google.android.gms.maps.model.SquareCap;
 import com.google.android.gms.maps.model.Tile;
 import com.google.maps.android.clustering.Cluster;
+import com.google.maps.android.heatmaps.Gradient;
+import com.google.maps.android.heatmaps.WeightedLatLng;
+import com.google.maps.android.projection.SphericalMercatorProjection;
 import io.flutter.FlutterInjector;
 import java.io.IOException;
 import java.io.InputStream;
@@ -39,6 +42,19 @@ import java.util.Map;
 
 /** Conversions between JSON-like values and GoogleMaps data types. */
 class Convert {
+  // These constants must match the corresponding constants in serialization.dart
+  public static final String HEATMAPS_TO_ADD_KEY = "heatmapsToAdd";
+  public static final String HEATMAPS_TO_CHANGE_KEY = "heatmapsToChange";
+  public static final String HEATMAP_IDS_TO_REMOVE_KEY = "heatmapIdsToRemove";
+  public static final String HEATMAP_ID_KEY = "heatmapId";
+  public static final String HEATMAP_DATA_KEY = "data";
+  public static final String HEATMAP_GRADIENT_KEY = "gradient";
+  public static final String HEATMAP_MAX_INTENSITY_KEY = "maxIntensity";
+  public static final String HEATMAP_OPACITY_KEY = "opacity";
+  public static final String HEATMAP_RADIUS_KEY = "radius";
+  public static final String HEATMAP_GRADIENT_COLORS_KEY = "colors";
+  public static final String HEATMAP_GRADIENT_START_POINTS_KEY = "startPoints";
+  public static final String HEATMAP_GRADIENT_COLOR_MAP_SIZE_KEY = "colorMapSize";
 
   private static BitmapDescriptor toBitmapDescriptor(
       Object o, AssetManager assetManager, float density) {
@@ -504,6 +520,19 @@ class Convert {
     return new LatLng(toDouble(data.get(0)), toDouble(data.get(1)));
   }
 
+  static WeightedLatLng toWeightedLatLng(Object o) {
+    final List<?> data = toList(o);
+    return new WeightedLatLng(toLatLng(data.get(0)), toDouble(data.get(1)));
+  }
+
+  private static final SphericalMercatorProjection sProjection = new SphericalMercatorProjection(1);
+
+  static Object weightedLatLngToJson(WeightedLatLng weightedLatLng) {
+    return Arrays.asList(
+        latLngToJson(sProjection.toLatLng(weightedLatLng.getPoint())),
+        weightedLatLng.getIntensity());
+  }
+
   static Point pointFromPigeon(Messages.PlatformPoint point) {
     return new Point(point.getX().intValue(), point.getY().intValue());
   }
@@ -885,8 +914,39 @@ class Convert {
     }
   }
 
+  /** Set the options in the given heatmap object to the given sink. */
+  static String interpretHeatmapOptions(Object o, HeatmapOptionsSink sink) {
+    final Map<?, ?> data = toMap(o);
+    final Object rawWeightedData = data.get(HEATMAP_DATA_KEY);
+    if (rawWeightedData != null) {
+      sink.setWeightedData(toWeightedData(rawWeightedData));
+    }
+    final Object gradient = data.get(HEATMAP_GRADIENT_KEY);
+    if (gradient != null) {
+      sink.setGradient(toGradient(gradient));
+    }
+    final Object maxIntensity = data.get(HEATMAP_MAX_INTENSITY_KEY);
+    if (maxIntensity != null) {
+      sink.setMaxIntensity(toDouble(maxIntensity));
+    }
+    final Object opacity = data.get(HEATMAP_OPACITY_KEY);
+    if (opacity != null) {
+      sink.setOpacity(toDouble(opacity));
+    }
+    final Object radius = data.get(HEATMAP_RADIUS_KEY);
+    if (radius != null) {
+      sink.setRadius(toInt(radius));
+    }
+    final String heatmapId = (String) data.get(HEATMAP_ID_KEY);
+    if (heatmapId == null) {
+      throw new IllegalArgumentException("heatmapId was null");
+    } else {
+      return heatmapId;
+    }
+  }
+
   @VisibleForTesting
-  static List<LatLng> toPoints(Object o) {
+  public static List<LatLng> toPoints(Object o) {
     final List<?> data = toList(o);
     final List<LatLng> points = new ArrayList<>(data.size());
 
@@ -895,6 +955,59 @@ class Convert {
       points.add(new LatLng(toDouble(point.get(0)), toDouble(point.get(1))));
     }
     return points;
+  }
+
+  /** Converts the given object to a list of WeightedLatLng objects. */
+  private static List<WeightedLatLng> toWeightedData(Object o) {
+    final List<?> data = toList(o);
+    final List<WeightedLatLng> weightedData = new ArrayList<>(data.size());
+
+    for (Object rawWeightedPoint : data) {
+      weightedData.add(toWeightedLatLng(rawWeightedPoint));
+    }
+    return weightedData;
+  }
+
+  /** Converts the given list of WeightedLatLng objects to a JSON object. */
+  public static Object weightedDataToJson(List<WeightedLatLng> weightedData) {
+    final List<Object> data = new ArrayList<>(weightedData.size());
+
+    for (WeightedLatLng weightedLatLng : weightedData) {
+      data.add(weightedLatLngToJson(weightedLatLng));
+    }
+    return data;
+  }
+
+  /** Converts the given object to a Gradient object. */
+  private static Gradient toGradient(Object o) {
+    final Map<?, ?> data = toMap(o);
+
+    final List<?> colorData = toList(data.get(HEATMAP_GRADIENT_COLORS_KEY));
+    assert colorData != null;
+    final int[] colors = new int[colorData.size()];
+    for (int i = 0; i < colorData.size(); i++) {
+      colors[i] = toInt(colorData.get(i));
+    }
+
+    final List<?> startPointData = toList(data.get(HEATMAP_GRADIENT_START_POINTS_KEY));
+    assert startPointData != null;
+    final float[] startPoints = new float[startPointData.size()];
+    for (int i = 0; i < startPointData.size(); i++) {
+      startPoints[i] = toFloat(startPointData.get(i));
+    }
+
+    final int colorMapSize = toInt(data.get(HEATMAP_GRADIENT_COLOR_MAP_SIZE_KEY));
+
+    return new Gradient(colors, startPoints, colorMapSize);
+  }
+
+  /** Converts the given Gradient object to a JSON object. */
+  private static Object gradientToJson(Gradient gradient) {
+    final Map<String, Object> data = new HashMap<>();
+    data.put(HEATMAP_GRADIENT_COLORS_KEY, gradient.mColors);
+    data.put(HEATMAP_GRADIENT_START_POINTS_KEY, gradient.mStartPoints);
+    data.put(HEATMAP_GRADIENT_COLOR_MAP_SIZE_KEY, gradient.mColorMapSize);
+    return data;
   }
 
   private static List<List<LatLng>> toHoles(Object o) {