+++ packages/google_maps_flutter/google_maps_flutter_android/android/src/test/java/io/flutter/plugins/googlemaps/GoogleMapControllerTest.java
@@ -53,6 +53,7 @@ public class GoogleMapControllerTest {
   @Mock PolygonsController mockPolygonsController;
   @Mock PolylinesController mockPolylinesController;
   @Mock CirclesController mockCirclesController;
+  @Mock HeatmapsController mockHeatmapsController;
   @Mock TileOverlaysController mockTileOverlaysController;
 
   @Before
@@ -86,6 +87,7 @@ public class GoogleMapControllerTest {
             mockPolygonsController,
             mockPolylinesController,
             mockCirclesController,
+            mockHeatmapsController,
             mockTileOverlaysController);
     googleMapController.init();
     return googleMapController;
@@ -222,5 +224,19 @@ public class GoogleMapControllerTest {
 
     googleMapController.onClusterItemClick(markerBuilder);
     verify(mockMarkersController, times(1)).onMarkerTap(markerBuilder.markerId());
+  }
+
+  @Test
+  public void SetInitialHeatmaps() {
+    GoogleMapController googleMapController = getGoogleMapControllerWithMockedDependencies();
+    Map<String, Object> initialHeatmap = new HashMap<>();
+    initialHeatmap.put("heatmapId", "hm_1");
+    List<Object> initialHeatmaps = new ArrayList<>();
+    initialHeatmaps.add(initialHeatmap);
+    googleMapController.setInitialHeatmaps(initialHeatmaps);
+    googleMapController.onMapReady(mockGoogleMap);
+
+    // Verify if the HeatmapsController.addHeatmaps method is called with initial heatmaps.
+    verify(mockHeatmapsController, times(1)).addHeatmaps(any());
   }
 }