+++ packages/google_maps_flutter/google_maps_flutter_web/example/integration_test/google_maps_plugin_test.mocks.dart
@@ -132,6 +132,7 @@ class MockGoogleMapController extends _i1.Mock
     _i4.DebugSetOptionsFunction? setOptions,
     _i4.MarkersController? markers,
     _i4.CirclesController? circles,
+    _i4.HeatmapsController? heatmaps,
     _i4.PolygonsController? polygons,
     _i4.PolylinesController? polylines,
     _i6.ClusterManagersController? clusterManagers,
@@ -146,6 +147,7 @@ class MockGoogleMapController extends _i1.Mock
             #setOptions: setOptions,
             #markers: markers,
             #circles: circles,
+            #heatmaps: heatmaps,
             #polygons: polygons,
             #polylines: polylines,
             #clusterManagers: clusterManagers,
@@ -270,6 +272,14 @@ class MockGoogleMapController extends _i1.Mock
   void updateCircles(_i2.CircleUpdates? updates) => super.noSuchMethod(
         Invocation.method(
           #updateCircles,
+          [updates],
+        ),
+        returnValueForMissingStub: null,
+      );
+  @override
+  void updateHeatmaps(_i2.HeatmapUpdates? updates) => super.noSuchMethod(
+        Invocation.method(
+          #updateHeatmaps,
           [updates],
         ),
         returnValueForMissingStub: null,